+++ src/main/api/ledger/evm/approve.ts
@@ -1,12 +1,13 @@
-import { FeeOption, TxHash } from '@xchainjs/xchain-client'
-import { ClientLedger } from '@xchainjs/xchain-evm'
+import TransportNodeHidSingleton from '@ledgerhq/hw-transport-node-hid-singleton'
+import { FeeOption, Network, TxHash } from '@xchainjs/xchain-client'
+import { ClientLedger, LedgerSigner } from '@xchainjs/xchain-evm'
 
 import { IPCLedgerApproveERC20TokenParams } from '../../../../shared/api/io'
 import { defaultArbParams } from '../../../../shared/arb/const'
 import { defaultAvaxParams } from '../../../../shared/avax/const'
 import { defaultBscParams } from '../../../../shared/bsc/const'
 import { defaultEthParams } from '../../../../shared/ethereum/const'
-import { getDerivationPaths } from '../../../../shared/evm/ledger'
+import { getDerivationPath, getDerivationPaths } from '../../../../shared/evm/ledger'
 
 export const approveLedgerERC20Token = async ({
   chain,
@@ -18,11 +19,16 @@ export const approveLedgerERC20Token = async ({
   hdMode
 }: IPCLedgerApproveERC20TokenParams): Promise<TxHash> => {
   let clientParams
-
+  const transport = await TransportNodeHidSingleton.create()
   switch (chain) {
     case 'ETH':
       clientParams = {
         ...defaultEthParams,
+        signer: new LedgerSigner({
+          transport,
+          provider: defaultEthParams.providers[Network.Mainnet],
+          derivationPath: getDerivationPath(walletAccount, hdMode)
+        }),
         rootDerivationPaths: getDerivationPaths(walletAccount, hdMode),
         network: network
       }
@@ -30,6 +36,11 @@ export const approveLedgerERC20Token = async ({
     case 'ARB':
       clientParams = {
         ...defaultArbParams,
+        signer: new LedgerSigner({
+          transport,
+          provider: defaultArbParams.providers[Network.Mainnet],
+          derivationPath: getDerivationPath(walletAccount, hdMode)
+        }),
         rootDerivationPaths: getDerivationPaths(walletAccount, hdMode),
         network: network
       }
@@ -37,6 +48,11 @@ export const approveLedgerERC20Token = async ({
     case 'AVAX':
       clientParams = {
         ...defaultAvaxParams,
+        signer: new LedgerSigner({
+          transport,
+          provider: defaultAvaxParams.providers[Network.Mainnet],
+          derivationPath: getDerivationPath(walletAccount, hdMode)
+        }),
         rootDerivationPaths: getDerivationPaths(walletAccount, hdMode),
         network: network
       }
@@ -44,6 +60,11 @@ export const approveLedgerERC20Token = async ({
     case 'BSC':
       clientParams = {
         ...defaultBscParams,
+        signer: new LedgerSigner({
+          transport,
+          provider: defaultBscParams.providers[Network.Mainnet],
+          derivationPath: getDerivationPath(walletAccount, hdMode)
+        }),
         rootDerivationPaths: getDerivationPaths(walletAccount, hdMode),
         network: network
       }